Removed the resize file
