--- src/graphics/resize.h
@@ -1,7 +1,0 @@
-#ifndef RESIZE_H
-#define RESIZE_H 
-
-void updateBounds(MenusBundle *menu_bundle, int *screen_width, int *screen_height);
-
-
-#endif
